refactored organization users list
